requirements.txt
webrequest==0.0.52
sqlalchemy_utils==0.33.11
mdx_linkify==1.2
apscheduler==3.5.3
psycopg2==2.7.7
babel==2.6.0
sqlalchemy_jsonfield==0.7.1
flask==1.0.2
alembic==1.0.6
flask-wtf==0.14.2
cherrypy==18.1.0
pyramid==1.10.1
beautifulsoup4==4.7.1
flask_sqlalchemy==2.3.2
flask-debugtoolbar==0.10.1

flask_caching==1.4.0 
pyinotify
python-Levenshtein
ftfy
colorama
paramiko
statsd
python-sql
irc
python-magic
git+https://github.com/fake-name/UniversalArchiveInterface.git
rpyc
sqlalchemy-citext
parsedatetime
cython
tqdm
natsort